terminando regra de negocio e criando 2 endpoints

+++ src/main/java/com/example/demo/dto/EstatisticaDto.java
@@ -1,8 +1,7 @@
 package com.example.demo.dto;
 
-
 public record EstatisticaDto(
-        int count,
+        long count,
         Double sum,
         Double avg,
         Double min,
